% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpiofind.R
\name{gpiofind}
\alias{gpiofind}
\title{Find the chip and line for a named function}
\usage{
gpiofind(gpio_name)
}
\arguments{
\item{gpio_name}{}
}
\value{
The chip name and line number.
}
\description{
Identifies the gpiochip and gpioline. For example, `gpiofind("SCL0")` will return "gpiochip0 45". Note that gpioset and gpioget both accept line names in place of line numbers and will identify the proper gpiochip using gpiofind behind the scenes.
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

gpiofind("SCL0") # returns "gpiochip0 45"

gpioget(gpiofind("SCL0")) # returns status of "gpiochip0 45"
\dontshow{\}) # examplesIf}
}
