% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpioget.R
\name{gpioget}
\alias{gpioget}
\title{gpioget}
\usage{
gpioget(gpio_chip = 0, gpio_active = TRUE, gpio_line, gpio_bias = "as-is")
}
\arguments{
\item{gpio_chip}{Retrieve the pin info from this chip number. Defaults to chip 0. Chips are listed by gpiodetect()}

\item{gpio_active}{Set the line active state, or what value to return when signal is high. gpio_active = FALSE inverts the active state of the line}

\item{gpio_line}{Retrieve state from this pin. Uses BCM numbering.}

\item{gpio_bias}{Set the line bias to one of the following: as-is (the default), disable, pull-down, pull-up}
}
\description{
gpioget
}
\examples{
read the value of a single gpio line
gpioget gpiochip1 23
Read two values at the same time. Set the active state of the lines to low.
gpioget --active-low gpiochip1 23 24
}
