% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rpi.R
\name{is.rpi}
\alias{is.rpi}
\title{DEPRECATED. Use rpi_whatami()$is_rpi instead.}
\usage{
is.rpi(whatami = rpi_whatami())
}
\arguments{
\item{whatami}{defaults to the data.frame returned by rpi_whatami()}
}
\value{
TRUE if running on a Raspberry Pi.
}
\description{
Confirm this code is running on a Raspberry Pi
}
\details{
Returns TRUE if this code is running on a Raspberry Pi
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
is.rpi()
\dontshow{\}) # examplesIf}
}
