% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rpi_spi.R
\name{is.rpi.spi}
\alias{is.rpi.spi}
\title{is SPI enabled on this Raspberry Pi?}
\usage{
is.rpi.spi()
}
\value{
TRUE if running on Raspberry Pi and SPI is enabled
}
\description{
is SPI enabled on this Raspberry Pi?
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
is.rpi.spi()
\dontshow{\}) # examplesIf}
}
