% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_i2c_dump.R
\name{rpi_i2c_dump}
\alias{rpi_i2c_dump}
\title{Read the entire data block from an I2C device}
\usage{
rpi_i2c_dump(chip_address)
}
\arguments{
\item{chip_address}{an integer between 0x03 and 0x77}
}
\value{
the entire memory block from the I2C device
}
\description{
This is a wrapper to i2cdump.
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rpi_i2c_dump(0x77)
\dontshow{\}) # examplesIf}
}
