% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_i2c_get.R
\name{rpi_i2c_get}
\alias{rpi_i2c_get}
\title{Return a byte from an i2c device}
\usage{
rpi_i2c_get(chip_address, data_address, data_size = "b")
}
\arguments{
\item{chip_address}{an integer between 0x03 and 0x77. Use i2cdetect -y 1 to identify the address}

\item{data_address}{an integer between 0x00 and 0xFF. This will come from the chip documentation}

\item{data_size}{read a byte (b) or a word (w) from data_address}
}
\value{
data from the i2c device located at chip_address and data_address
}
\description{
Return a byte from an i2c device
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rpi_i2c_get(chip_address = 0x77, data_address = 0xfa, data_size = 'b') # get a byte from an i2c device located at 0x77
\dontshow{\}) # examplesIf}
}
