% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_i2c_set.R
\name{rpi_i2c_set}
\alias{rpi_i2c_set}
\title{Write data to an i2c device}
\usage{
rpi_i2c_set(chip_address, data_address, value, data_size = "b")
}
\arguments{
\item{chip_address}{an integer between 0x03 and 0x77. Use i2cdetect -y 1 to identify the address}

\item{data_address}{an integer between 0x00 and 0xFF. This will come from the chip documentation}

\item{value}{a byte (0xff), a word (0xffff), a SMBus (32 bytes) or I2C Block}

\item{data_size}{}
}
\value{
void
}
\description{
Writes a byte, a word, or other data block to an i2c device.
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
