% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_monitor.R
\name{rpi_monitor}
\alias{rpi_monitor}
\title{Monitor an RPi pin}
\usage{
rpi_monitor(pin_number, numEvents = 0, edge = "both")
}
\arguments{
\item{pin_number}{one or more pin numbers as found on the Raspberry pi GPIO. Use the actual pin numbers (i.e. 1:40)}

\item{numEvents}{a number of events to collect. Defaults to 10}

\item{edge}{rising|falling|both}
}
\value{
a list of events
\itemize{
\item raw text of each event
\item event: RISING EDGE|FALLING EDGE
\item line: the bcm line (not board pin) exhibiting change
\item timestamp: seconds component
\item timestamp: nanoseconds component
}
}
\description{
Watch a Raspberry pi GPIO pin (or pins) for changes.
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
rpi_monitor(21)
rpi_monitor(21, numEvents = 20)
}
\dontshow{\}) # examplesIf}
}
