% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_pinToBCM.R
\name{rpi_pinToBCM}
\alias{rpi_pinToBCM}
\title{Converts a RPi pin number to BCM line}
\usage{
rpi_pinToBCM(pin_number)
}
\arguments{
\item{pin_number}{The pin as found on the Raspberry Pi board. Numbered from 1 to 40.}
}
\value{
a named vector.
bcm_line = broadcom line connected to this pin
pin_name = broadcom name such as RXD1 or GPIO21
}
\description{
Look up a board pin number and return BCM_line and line_name
Pin number is the physical board of the Raspberry pi.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This checks for a valid range of pins (1:40) and for valid input/output lines. This prevents trying to read a 3.3v or ground line.
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
