% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_pwm.R
\name{rpi_pwm}
\alias{rpi_pwm}
\title{Provide R with access to PWM}
\usage{
rpi_pwm(pin_number = 12, mode = "help")
}
\arguments{
\item{pin_number.}{Select one or two of four hardware PWM lines (12, 32, 33, or 35). Generates an error message if a pwm channel is selected twice.
Since PWM0 appears on both pins 12 and 32, and PWM1 appears on pins 33 and 35, a valid selection of two pins is one from PWM0 and one from PWM1.
In other words, only these combinations are valid:
\itemize{
   \item pin_number = c(12,33)
   \item pin_number = c(32,33)
   \item pin_number = c(12,35)
   \item pin_number = c(32,35)
}}

\item{mode.}{An activity related to PWM

    \itemize{
     \item help. Prints an informative message
     \item enable. Prints instructions for enabling the selected PWM pin. This requires root access which R doesn't have
   }}
}
\value{
void
}
\description{
Provides the R programming language with access to Raspberry Pi hardware Pulse Width Modulation (PWM)
}
\details{
Only two PWM lines are available on the Raspberry Pi - pwm0 and pwm1. Although it is possible to create software PWM, this isn't advised due to latency of linux and R.
    It might be helpful to read the article on `rpi_pwm()` located at the [rpigpior website](https://mnr.github.io/rpigpior/articles/rpi_pwm.html).
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

\dontshow{\}) # examplesIf}
}
