% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpi_spi_open}
\alias{rpi_spi_open}
\title{Open an SPI session}
\usage{
rpi_spi_open(spiChan, spiBus = 0L, max_speed_hz = 32000000L)
}
\arguments{
\item{spiChan}{The SPI channel to use. /dev/spidev[spiBus].[spiChan]}

\item{spiBus}{The SPI Bus to use, defaults to bus 0.}

\item{max_speed_hz}{spi transfer speed}
}
\value{
SPI identifier
}
\description{
This assumes the SPI device is attached to bus zero.
}
\examples{
\dontshow{if (is.rpi.spi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
spi_identifier <- rpi_spi_open(spiBus = 0, spiChan = 0)
\dontshow{\}) # examplesIf}
}
