% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_whatami.R
\name{rpi_whatami}
\alias{rpi_whatami}
\title{rpi_whatami}
\usage{
rpi_whatami()
}
\value{
A data.frame with information about this RPi

If this isn't a Raspberry Pi, ...$is_rpi will be FALSE. In addition, ...$fault will contain the error message

If this is a Raspberry Pi, the data frame will contain the following:
\itemize{
\item ...$is_rpi TRUE if this is a Raspberry Pi
\item ...$rpi_version "Zero", 1, 2, 3, 4, or 5
\item ...$gpiochip normally gpiochip0. For RPi 5, returns gpiochip4
\item ...$i2c_enabled TRUE if i2c is enabled on this RPi
\item ...$spi_enabled TRUE if spi is enabled on this RPi
\item ...$gpiod_enabled TRUE if gpiod is installed on this RPi
}
}
\description{
Identify the model of Raspberry Pi this code is running on
including useful environmental information
}
\examples{
\dontshow{if (rpigpior::rpi_whatami()$is_rpi) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rpigpior::rpi_whatami()
\dontshow{\}) # examplesIf}
}
