% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpigpior-package.R
\docType{package}
\name{rpigpior-package}
\alias{rpigpior}
\alias{rpigpior-package}
\title{rpigpior: Provides R with Access to the Raspberry Pi}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This provides R with access to the Raspberry Pi General Purpose Input Output (GPIO) including GPIO pins, i2c, and SPI. No additional libraries need to be installed; Raspbian includes 'libgpiod' as part of the standard installation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mnr/rpigpior}
  \item \url{https://mnr.github.io/rpigpior/}
  \item Report bugs at \url{https://github.com/mnr/rpigpior/issues}
}

}
\author{
\strong{Maintainer}: Mark Niemann-Ross \email{mark@niemannross.com} (\href{https://orcid.org/0000-0002-3381-4604}{ORCID}) [copyright holder]

}
\keyword{internal}
