% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spiClose}
\alias{spiClose}
\title{Close an SPI session}
\usage{
spiClose(spiDeviceID)
}
\arguments{
\item{spiDeviceID}{An identifier supplied by rpi_spi_open()}
}
\value{
SPI identifier
}
\description{
Close an SPI session
}
