% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpiodetect.R
\name{gpiodetect}
\alias{gpiodetect}
\title{Find gpiochips on this system}
\usage{
gpiodetect(gpio_chip = "")
}
\arguments{
\item{gpio_chip}{}
}
\value{
a (list? data.frame? vector?) with name, label, and number of GPIO lines
}
\description{
Find gpiochips on this system
}
\examples{
gpiodetect()
gpiodetect(0)
}
