% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpiodetect.R
\name{gpiodetect}
\alias{gpiodetect}
\title{Find gpiochips on this system}
\usage{
gpiodetect()
}
\value{
a chr vector with name, label, and number of GPIO lines. Typically this will return something like the following:

    gpiochip0 [pinctrl-bcm2711] (58 lines)

    gpiochip1 [raspberrypi-exp-gpio] (8 lines)
}
\description{
Find gpiochips on this system
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gpiodetect()
\dontshow{\}) # examplesIf}
}
