% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpiodetect.R
\name{gpiodetect}
\alias{gpiodetect}
\title{Find gpiochips on this system}
\usage{
gpiodetect()
}
\value{
a chr vector with name, label, and number of GPIO lines
}
\description{
Find gpiochips on this system
}
\examples{
gpiodetect()
}
