% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpioget.R
\name{gpioget}
\alias{gpioget}
\title{gpioget}
\usage{
gpioget(gpio_chip, gpio_active = FALSE, gpio_line)
}
\arguments{
\item{gpio_active}{}
}
\description{
gpioget
}
\examples{
read the value of a single gpio line
gpioget gpiochip1 23
Read two values at the same time. Set the active state of the lines to low.
gpioget --active-low gpiochip1 23 24
}
