% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rpi.R
\name{is.rpi}
\alias{is.rpi}
\title{Confirm this code is running on a Raspberry Pi}
\usage{
is.rpi(tellme = FALSE)
}
\arguments{
\item{tellme}{if TRUE the is.rpi will return the device-tree/model.
This typically looks like "Raspberry Pi 5 Model B Rev 1.0"}
}
\value{
TRUE if running on a Raspberry Pi.
if tellme = TRUE, return the model string
}
\description{
Returns TRUE if this code is running on a Raspberry Pi
}
\examples{
\dontshow{if (is.rpi()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
is.rpi()
\dontshow{\}) # examplesIf}
}
