% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_set.R
\name{rpi_set}
\alias{rpi_set}
\title{Turn lines on or off}
\usage{
rpi_set(pin_number, onOff)
}
\arguments{
\item{pin_number}{one or more pin numbers as found on the Raspberry pi GPIO. Use the actual pin numbers (i.e. 1:40)}

\item{onOff}{a vector the same length of pin_number with values of 0 or 1 to turn the line off or on}
}
\value{
no value is returned
}
\description{
Turn lines on or off
}
\examples{
rpi_set(pin_number = c(9,10,11), onOff = c(1,0,1))
}
